Log Management/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.9
redis==5.0.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
pydantic==2.10.4
pydantic-settings==2.8.0
loguru==0.7.2
cryptography==42.0.8
paramiko==3.4.0
ldap3==2.9.1
requests==2.32.3
pandas==2.2.3
openpyxl==3.1.2
python-dateutil==2.9.0
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.0 
PyJWT==2.8.0
geoip2==4.8.0
aiohttp==3.10.11
pysnmp==6.2.6
clickhouse-connect==0.7.16
openai==1.54.5
langchain==0.3.7
tiktoken==0.8.0 